Changed C wrapper. RetroCompatibility available
