+++ examples/c_zhtclient_main.c
@@ -3,11 +3,13 @@
 #include   <stdio.h>
 
 #include   <string.h>
-#include "c_zhtclient.h"
+#include <c_zhtclientStd.h>
 
 const int LOOKUP_SIZE = 65535;
 
 int main(int argc, char **argv) {
+
+	ZHTClient_c zhtClient;
 
 	if (argc < 4) {
 
@@ -25,7 +27,7 @@ int main(int argc, char **argv) {
 		useTCP = false;
 	}
 
-	c_zht_init(argv[1], argv[2], useTCP); //neighbor zht.cfg TCP
+	c_zht_init_std(&zhtClient, argv[1], argv[2], useTCP); //neighbor zht.cfg TCP
 
 	const char *key = "hello";
 	const char *value = "zht";
@@ -36,23 +38,23 @@ int main(int argc, char **argv) {
 	key = largeKey;
 	value = largeVal;
 
-	int iret = c_zht_insert2(key, value);
+	int iret = c_zht_insert2_std(zhtClient, key, value);
 	fprintf(stderr, "c_zht_insert, return code: %d\n", iret);
 
 	size_t n;
 	char *result = (char*) calloc(LOOKUP_SIZE, sizeof(char));
 	if (result != NULL) {
-		int lret = c_zht_lookup2(key, result, &n);
+		int lret = c_zht_lookup2_std(zhtClient, key, result, &n);
 		fprintf(stderr, "c_zht_lookup, return code: %d\n", lret);
 		fprintf(stderr, "c_zht_lookup, return value: length(%lu), %s\n", n,
 				result);
 	}
 	free(result);
 
-	int rret = c_zht_remove2(key);
+	int rret = c_zht_remove2_std(zhtClient, key);
 	fprintf(stderr, "c_zht_remove, return code: %d\n", rret);
 
-	c_zht_teardown();
+	c_zht_teardown_std(zhtClient);
 
 	return 0;
 }
